Refactor top-level app to make the runtime testable
